+++ src/app/admin/admin-management/admin-management.component.ts
@@ -2,8 +2,9 @@ import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { Subscription } from 'rxjs';
 import { AuthService } from 'src/app/security/auth.service';
-import { User } from 'src/app/security/user';
-import { UserService } from 'src/app/security/user.service';
+import { User } from 'src/app/shared/services/user';
+
+import { UserService } from 'src/app/shared/services/user.service';
 
 @Component({
   selector: 'app-admin-management',
