+++ src/app/admin/product/product-overview/product-overview.component.ts
@@ -3,7 +3,7 @@ import { Router } from '@angular/router';
 import { Product } from 'src/app/admin/product/product';
 import { ProductService } from '../product.service';
 import { AuthService } from 'src/app/security/auth.service';
-import { UserService } from 'src/app/security/user.service';
+import { UserService } from 'src/app/shared/services/user.service';
 
 
 @Component({
