+++ src/app/homepage/product-detail/product-detail.component.ts
@@ -4,7 +4,7 @@ import { ActivatedRoute } from '@angular/router';
 import { Subscription } from 'rxjs';
 import { Category } from '../../admin/category/category';
 
-import { Order } from '../../user/account/order';
+import { Order } from '../../shared/services/order';
 import { Product } from '../../admin/product/product';
 
 import { AuthService } from '../../security/auth.service';
@@ -13,7 +13,8 @@ import { Basket } from 'src/app/user/basket/basket';
 import { BasketItem } from 'src/app/user/basket/basketItem';
 import { BasketService } from 'src/app/user/basket/basket.service';
 import { BasketItemService } from 'src/app/user/basket/basket-item.service';
-import { User } from 'src/app/security/user';
+import { User } from 'src/app/shared/services/user';
+
 
 @Component({
   selector: 'app-product-detail',
