+++ src/app/security/security.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { SecurityComponent } from './security/security.component';
 import { FormsModule } from '@angular/forms';
-import { UserService } from './user.service';
+import { UserService } from '../shared/services/user.service';
