+++ src/app/security/security/security.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit } from '@angular/core';
-import { User } from '../user';
 import { AuthService } from '../auth.service';
 import { Router } from '@angular/router';
 import emailjs, { EmailJSResponseStatus } from 'emailjs-com';
+import { User } from 'src/app/shared/services/user';
 
 @Component({
   selector: 'app-security',
