+++ src/app/security/userResponse.ts
@@ -1,5 +1,3 @@
-import { User } from './user';
-
 export interface UserResponse {
   user_id: string;
   access_token: string;
