+++ src/app/shared/services/services.module.ts
@@ -1,0 +1,18 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { UserService } from './user.service';
+import { OrderService } from './order.service';
+
+
+
+@NgModule({
+  declarations: [],
+  imports: [
+    CommonModule
+  ],
+  providers: [
+    UserService,
+    OrderService
+  ]
+})
+export class ServicesModule { }
