+++ src/app/user/account/account-overview/account-overview.component.ts
@@ -1,8 +1,9 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { Subscription } from 'rxjs';
+import { User } from 'src/app/shared/services/user';
 import { AuthService } from '../../../security/auth.service';
-import { User } from '../../../security/user';
-import { UserService } from '../../../security/user.service';
+
+import { UserService } from '../../../shared/services/user.service';
 import { Basket } from '../../basket/basket';
 import { BasketService } from '../../basket/basket.service';
