+++ src/app/user/account/account.module.ts
@@ -4,7 +4,7 @@ import { AccountFormComponent } from './account-form/account-form.component';
 import { AccountOverviewComponent } from './account-overview/account-overview.component';
 import { AccountOrderComponent } from './account-order/account-order.component';
 import { SharedModule } from 'src/app/shared/shared.module';
-import { OrderService } from './order.service';
+import { OrderService } from '../../shared/services/order.service';
