+++ src/app/user/basket/basket.ts
@@ -1,6 +1,5 @@
-import { Order } from "src/app/user/account/order";
-import { User } from "src/app/security/user";
-
+import { Order } from "src/app/shared/services/order";
+import { User } from "src/app/shared/services/user";
 
 export interface Basket {
   _id: string;
